LPD-88102 Unify export and import preview display contexts
