+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/import/new_import.jsp
@@ -10,37 +10,15 @@
 <liferay-staging:defineObjects />
 
 <%
-String backURL = ParamUtil.getString(request, "backURL");
-
 if (liveGroup == null) {
 	liveGroup = group;
 	liveGroupId = groupId;
 }
 
-ExportImportPreviewDisplayContext exportImportPreviewDisplayContext = new ExportImportPreviewDisplayContext(liveGroup);
-
-String displayStyle = ParamUtil.getString(request, "displayStyle");
-
-PortletURL portletURL = PortletURLBuilder.createRenderURL(
-	renderResponse
-).setMVCRenderCommandName(
-	"/export_import/view_import_layouts"
-).setParameter(
-	"displayStyle", displayStyle
-).setParameter(
-	"groupId", groupId
-).setParameter(
-	"liveGroupId", liveGroupId
-).setParameter(
-	"privateLayout", privateLayout
-).buildPortletURL();
-
-if (Validator.isBlank(backURL)) {
-	backURL = portletURL.toString();
-}
+ExportImportPreviewDisplayContext exportImportPreviewDisplayContext = new ExportImportPreviewDisplayContext("/export_import/view_import_layouts", request, liferayPortletResponse, liveGroup, groupId, liveGroupId, privateLayout, stagingGroupHelper);
 
 portletDisplay.setShowBackIcon(true);
-portletDisplay.setURLBack(backURL);
+portletDisplay.setURLBack(exportImportPreviewDisplayContext.getBackURL());
 %>
 
 <clay:container-fluid
@@ -54,7 +32,7 @@ portletDisplay.setURLBack(backURL);
 		module="{NewImport} from exportimport-web"
 		props='<%=
 			HashMapBuilder.<String, Object>put(
-				"backURL", backURL
+				"backURL", exportImportPreviewDisplayContext.getBackURL()
 			).put(
 				"importPreviewAPIURL", exportImportPreviewDisplayContext.getImportPreviewAPIURL()
 			).build()
