+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/init.jsp
@@ -6,7 +6,6 @@
 --%>
 
 <%@ page import="com.liferay.exportimport.web.internal.display.context.ExportImportPreviewDisplayContext" %><%@
-<%@ page import="com.liferay.exportimport.web.internal.display.context.NewExportDisplayContext" %><%@
 page import="com.liferay.portal.kernel.json.JSONArray" %><%@
 page import="com.liferay.portal.kernel.json.JSONFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.util.PropsValues" %>
